@@ -1,46 +1,65 @@
 {
   "scripts": {
-    "dev": "npm run build-dev & $(npm bin)/nodemon build/index.js",
+    "dev": "nodemon build/index.js",
     "clean-build": "$(npm bin)/rimraf build",
-    "build-dev": "$(npm bin)/webpack-dev-server --no-inline --progress --port 3000",
+    "build-dev": "webpack serve --progress --port 3000",
     "build-prod": "$(npm bin)/webpack --config webpack/webpack.prod.ts",
+    "create-products": "ts-node seed/products.ts",
+    "lint": "eslint \"**/*.ts\"",
+    "lint:fix": "npm run lint -- --fix",
+    "format": "prettier -l \"**/*.ts\"",
+    "format:fix": "npm run format -- --write",
     "test": "NODE_ENV=unittest mocha -r ts-node/register --harmony $(find test/unit -name *.test.ts)",
     "test-sample": "mocha -r ts-node/register test/*.test.ts"
   },
   "dependencies": {
     "@types/chalk": "2.2.0",
     "bcryptjs": "2.4.3",
-    "chalk": "2.4.0",
-    "compression": "1.7.2",
-    "cors": "2.8.4",
-    "express": "4.16.3",
-    "express-validator": "5.2.0",
-    "inversify": "4.13.0",
-    "jwt-simple": "0.5.1",
-    "method-override": "2.3.10",
-    "moment": "2.22.1",
-    "mongoose": "5.0.16",
-    "morgan": "1.9.0",
-    "nodemon": "1.17.4",
-    "reflect-metadata": "0.1.12",
-    "typescript": "2.8.3",
-    "winston": "3.0.0-rc5"
+    "chalk": "4.1.0",
+    "compression": "1.7.4",
+    "cors": "2.8.5",
+    "ejs": "^3.1.6",
+    "express": "4.17.1",
+    "express-validator": "6.10.0",
+    "inversify": "5.0.5",
+    "jwt-simple": "0.5.6",
+    "method-override": "3.0.0",
+    "moment": "2.29.1",
+    "mongoose": "5.11.18",
+    "morgan": "1.10.0",
+    "nodemon": "2.0.7",
+    "reflect-metadata": "0.1.13",
+    "typescript": "4.2.2",
+    "winston": "3.3.3"
   },
   "devDependencies": {
-    "@types/chai": "4.1.3",
-    "@types/mocha": "5.2.1",
-    "@types/mongoose": "5.0.10",
-    "@types/node": "9.6.6",
-    "chai": "^4.1.2",
-    "mocha": "5.2.0",
-    "rimraf": "^2.6.2",
-    "sinon": "^5.1.1",
-    "supertest": "^3.1.0",
-    "ts-loader": "4.2.0",
-    "ts-node": "6.1.0",
-    "webpack": "4.6.0",
-    "webpack-cli": "2.0.15",
-    "webpack-dev-server": "3.1.3",
-    "webpack-merge": "4.1.2"
+    "@types/bcryptjs": "2.4.2",
+    "@types/body-parser": "1.19.0",
+    "@types/chai": "4.2.15",
+    "@types/compression": "1.7.0",
+    "@types/cors": "2.8.10",
+    "@types/ejs": "3.0.6",
+    "@types/express": "4.17.11",
+    "@types/method-override": "0.0.31",
+    "@types/mocha": "8.2.1",
+    "@types/mongoose": "5.10.3",
+    "@types/morgan": "1.9.2",
+    "@types/node": "14.14.31",
+    "@typescript-eslint/eslint-plugin": "4.15.2",
+    "@typescript-eslint/parser": "4.15.2",
+    "chai": "4.3.0",
+    "eslint": "7.21.0",
+    "eslint-config-prettier": "8.1.0",
+    "mocha": "8.3.0",
+    "prettier": "^2.2.1",
+    "rimraf": "3.0.2",
+    "sinon": "9.2.4",
+    "supertest": "6.1.3",
+    "ts-loader": "8.0.17",
+    "ts-node": "9.1.1",
+    "webpack": "5.24.2",
+    "webpack-cli": "4.5.0",
+    "webpack-dev-server": "3.11.2",
+    "webpack-merge": "5.7.3"
   }
 }
