@@ -1,9 +1,8 @@
-let path = require('path');
-let rootPath = path.resolve(__dirname, '..');
+import * as path from 'path';
+const rootPath = path.resolve(__dirname, '..');
 
-function root (args) {
-  args = Array.prototype.slice.call(arguments, 0);
-  return path.join.apply(path, [rootPath].concat(args));
+function root(...args: any) {
+  return path.join(...[rootPath].concat(args));
 }
 
 exports.root = root;
