@@ -1,37 +1,48 @@
-import {
-  ApiFactory
-} from './api/api-factory';
-import {
-  Server
-} from './server';
-import {
-  Container
-} from 'inversify';
-import {
-  Injectable
-} from './util-lib/injectable';
-import TYPES from './constant/types';
+import { ApiFactory } from './api/api-factory';
 import * as express from 'express';
-import * as expressValidator from 'express-validator';
+import { ExpressConfig } from './config/express';
+import { ResponseData } from './util-lib/response-data';
+import { Logger } from './util-lib/logger';
+import { ErrorHandler } from './util-lib/error-handler';
 
 export class Api {
-  constructor () {
-    this.injectableContainer.bind<Injectable>(TYPES.Injectable).to(Injectable);
-    this.injectable = this.injectableContainer.get<Injectable>(TYPES.Injectable);
-    this.app = this.injectable.server.getApp();
-    this.loadRoutes();
-  }
-
+  public app: express.Application = express();
+  private expressConfig: ExpressConfig = new ExpressConfig(this.app);
   private static instance: Api = new Api();
-  private injectableContainer = new Container();
-  private injectable: Injectable;
-  private app: any;
+
+  public onInit(): void {
+    this.expressConfig.bindConfig();
+    this.loadRoutes();
+    this.loadErrorHandler();
+  }
 
   private loadRoutes() {
+    this.app.all(
+      '*',
+      (_request: express.Request, _response: express.Response, next: express.NextFunction) => {
+        next();
+      }
+    );
+
     ApiFactory.createAllRoutes(this.app);
+
+    this.app.get(
+      '*',
+      (req: express.Request, res: express.Response, _next: express.NextFunction) => {
+        res.statusCode = 404;
+        res.json(ResponseData.create404('Unable to find the requested path.'));
+        Logger.error(
+          (`An unregistered path '${req.originalUrl}' was requested` as unknown) as Error
+        );
+      }
+    );
   }
 
-  public static getApp() {
-    return Api.instance.app;
+  private loadErrorHandler() {
+    this.app.use(ErrorHandler.init());
+  }
+
+  public static getApi(): Api {
+    return Api.instance;
   }
 }
