@@ -1,23 +1,25 @@
-import {
-  UserGetApi
-} from './user/get';
 import * as express from 'express';
+import { ProductGetApi } from './product/get';
+import { CartPostApi } from './cart/post';
 
 export class ApiFactory {
-  constructor() {}
-
   private static instance: ApiFactory = new ApiFactory();
   private app: express.Application;
-  private URLPrefixed: string = '/api/v1';
+  private URLPrefixed = '/api/v1';
 
-  public static createAllRoutes(app: express.Application) {
+  public static createAllRoutes(app: express.Application): void {
     this.instance.app = app;
-    this.instance.createUserRouter()
+    this.instance.createProductRouter();
+    this.instance.createCartRouter();
   }
 
-  public createUserRouter() {
-    this.app.use(this.URLPrefixed, UserGetApi.getInstance());
+  public createProductRouter(): this {
+    this.app.use(this.URLPrefixed, ProductGetApi.getInstance());
 
     return this;
+  }
+
+  public createCartRouter(): void {
+    this.app.use(this.URLPrefixed, CartPostApi.getInstance());
   }
 }
