@@ -1,8 +1,10 @@
 const TYPES = {
   Server: Symbol.for('Server'),
   ExpressConfig: Symbol.for('ExpressConfig'),
+  ProductModel: Symbol.for('ProductModel'),
+  CartModel: Symbol.for('CartModel'),
   Modules: Symbol.for('Modules'),
-  Injectable: Symbol.for('Modules')
+  Injectable: Symbol.for('Injectable'),
 };
 
 export default TYPES;
