@@ -1,23 +1,12 @@
 import TYPES from '../constant/types';
-import {
-  Container,
-  injectable
-} from 'inversify';
-import {
-  Server
-} from '../server';
+import { Container } from 'inversify';
+import { Server } from '../server';
+import { ProductModel } from '../model/product';
+import { CartModel } from '../model/cart';
 
-@injectable()
-export class Injectable {
-  constructor() {
-    this.onInit();
-  }
+const container = new Container();
+container.bind<ProductModel>(TYPES.ProductModel).to(ProductModel);
+container.bind<CartModel>(TYPES.CartModel).to(CartModel);
+container.bind<Server>(TYPES.Server).to(Server);
 
-  private serverContainer = new Container();
-  public server: Server;
-
-  private onInit(): void {
-    this.serverContainer.bind<Server>(TYPES.Server).to(Server);
-    this.server = this.serverContainer.get<Server>(TYPES.Server);
-  }
-}
+export { container };
